+++ lib/ts/recipe/emailpassword/components/library/input.tsx
@@ -20,7 +20,6 @@ import CheckedIcon from "../../../../components/assets/checkedIcon";
 import ErrorIcon from "../../../../components/assets/errorIcon";
 import ShowPasswordIcon from "../../../../components/assets/showPasswordIcon";
 
-import type { APIFormField } from "../../../../types";
 import type { ChangeEvent } from "react";
 
 export type InputProps = {
@@ -32,9 +31,9 @@ export type InputProps = {
     hasError: boolean;
     placeholder: string;
     value: string;
-    onInputBlur?: (field: APIFormField) => void;
-    onInputFocus?: (field: APIFormField) => void;
-    onChange?: (field: APIFormField) => void;
+    onInputBlur?: (value: string) => void;
+    onInputFocus?: (value: string) => void;
+    onChange?: (value: string) => void;
 };
 
 const Input: React.FC<InputProps> = ({
@@ -59,28 +58,19 @@ const Input: React.FC<InputProps> = ({
 
     function handleFocus() {
         if (onInputFocus !== undefined) {
-            onInputFocus({
-                id: name,
-                value: value,
-            });
+            onInputFocus(value);
         }
     }
 
     function handleBlur() {
         if (onInputBlur !== undefined) {
-            onInputBlur({
-                id: name,
-                value,
-            });
+            onInputBlur(value);
         }
     }
 
     function handleChange(event: ChangeEvent<HTMLInputElement>) {
         if (onChange) {
-            onChange({
-                id: name,
-                value: event.target.value,
-            });
+            onChange(event.target.value);
         }
     }