+++ test/end-to-end/signup.test.js
@@ -39,6 +39,8 @@ import {
     getGeneralError,
     waitForSTElement,
     backendBeforeEach,
+    setSelectDropdownValue,
+    getInputField,
 } from "../helpers";
 
 import {
@@ -340,6 +342,314 @@ describe("SuperTokens SignUp", function () {
 
             let [emailError] = await getFieldErrors(page);
             assert.deepStrictEqual(emailError, "This email already exists. Please sign in instead.");
+        });
+    });
+
+    describe("Signup custom fields test", function () {
+        beforeEach(async function () {
+            // set cookie and reload which loads the form with custom field
+            await page.evaluate(() => window.localStorage.setItem("SHOW_CUSTOM_FIELDS", "YES"));
+
+            await page.reload({
+                waitUntil: "domcontentloaded",
+            });
+            await toggleSignInSignUp(page);
+        });
+
+        it("Check if the custom fields are loaded", async function () {
+            let text = await getAuthPageHeaderText(page);
+            assert.deepStrictEqual(text, "Sign Up");
+
+            // check if select dropdown is loaded
+            const selectDropdownExists = await waitForSTElement(page, "select");
+            assert.ok(selectDropdownExists, "Select dropdown exists");
+
+            // check if checbox is loaded
+            const checkboxExists = await waitForSTElement(page, 'input[type="checkbox"]');
+            assert.ok(checkboxExists, "Checkbox exists");
+        });
+
+        it("Should show error messages, based on optional flag", async function () {
+            await submitForm(page);
+            let formFieldErrors = await getFieldErrors(page);
+
+            // 2 regular form field errors +
+            // 1 required custom field => terms checkbox
+            assert.deepStrictEqual(formFieldErrors, [
+                "Field is not optional",
+                "Field is not optional",
+                "Field is not optional",
+            ]);
+
+            // supply values for regular-required fields only
+            await setInputValues(page, [
+                { name: "email", value: "jack.doe@supertokens.io" },
+                { name: "password", value: "Str0ngP@ssw0rd" },
+            ]);
+
+            await submitForm(page);
+            formFieldErrors = await getFieldErrors(page);
+            assert.deepStrictEqual(formFieldErrors, ["Field is not optional"]);
+
+            // check terms and condition checkbox
+            let termsCheckbox = await waitForSTElement(page, '[name="terms"]');
+            await page.evaluate((e) => e.click(), termsCheckbox);
+
+            //un-checking the required checkbox should throw custom error message
+            await page.evaluate((e) => e.click(), termsCheckbox);
+
+            await submitForm(page);
+            formFieldErrors = await getFieldErrors(page);
+            assert.deepStrictEqual(formFieldErrors, ["Please check Terms and conditions"]);
+        });
+
+        it("Check if custom values are part of the signup payload", async function () {
+            const customFields = {
+                terms: "true",
+                "select-dropdown": "option 3",
+            };
+            let assertionError = null;
+            let interceptionPassed = false;
+
+            const requestHandler = async (request) => {
+                if (request.url().includes(SIGN_UP_API) && request.method() === "POST") {
+                    try {
+                        const postData = JSON.parse(request.postData());
+                        Object.keys(customFields).forEach((key) => {
+                            let findFormData = postData.formFields.find((inputData) => inputData.id === key);
+                            if (findFormData) {
+                                assert.strictEqual(
+                                    findFormData["value"],
+                                    customFields[key],
+                                    `Mismatch in payload for key: ${key}`
+                                );
+                            } else {
+                                throw new Error("Field not found in req.data");
+                            }
+                        });
+                        interceptionPassed = true;
+                        return request.respond({
+                            status: 200,
+                            headers: {
+                                "access-control-allow-origin": TEST_CLIENT_BASE_URL,
+                                "access-control-allow-credentials": "true",
+                            },
+                            body: JSON.stringify({
+                                status: "OK",
+                            }),
+                        });
+                    } catch (error) {
+                        assertionError = error; // Store the error
+                    }
+                }
+                return request.continue();
+            };
+
+            await page.setRequestInterception(true);
+            page.on("request", requestHandler);
+
+            try {
+                // Fill and submit the form with custom fields
+                await setInputValues(page, [
+                    { name: "email", value: "john.doe@supertokens.io" },
+                    { name: "password", value: "Str0ngP@assw0rd" },
+                ]);
+
+                await setSelectDropdownValue(page, "select", customFields["select-dropdown"]);
+
+                // Check terms and condition checkbox
+                let termsCheckbox = await waitForSTElement(page, '[name="terms"]');
+                await page.evaluate((e) => e.click(), termsCheckbox);
+
+                // Perform the button click and wait for all network activity to finish
+                await Promise.all([page.waitForNetworkIdle(), submitForm(page)]);
+            } finally {
+                page.off("request", requestHandler);
+                await page.setRequestInterception(false);
+            }
+
+            if (assertionError) {
+                throw assertionError;
+            }
+
+            if (!interceptionPassed) {
+                throw new Error("test failed");
+            }
+        });
+    });
+
+    // Default values test
+    describe("Signup default value for fields test", function () {
+        beforeEach(async function () {
+            // set cookie and reload which loads the form fields with default values
+            await page.evaluate(() => window.localStorage.setItem("SHOW_CUSTOM_FIELDS_WITH_DEFAULT_VALUES", "YES"));
+
+            await page.reload({
+                waitUntil: "domcontentloaded",
+            });
+            await toggleSignInSignUp(page);
+        });
+
+        it("Check if default values are set already", async function () {
+            const fieldsWithDefault = {
+                country: "India",
+                "select-dropdown": "option 2",
+                terms: true,
+            };
+
+            // regular input field default value
+            const countryInput = await getInputField(page, "country");
+            const defaultCountry = await countryInput.evaluate((f) => f.value);
+            assert.strictEqual(defaultCountry, fieldsWithDefault["country"]);
+
+            // custom dropdown default value is also getting set correctly
+            const selectDropdown = await waitForSTElement(page, "select");
+            const defaultOption = await selectDropdown.evaluate((f) => f.value);
+            assert.strictEqual(defaultOption, fieldsWithDefault["select-dropdown"]);
+
+            // custom dropdown default value is also getting set correctly
+            const termsCheckbox = await waitForSTElement(page, '[name="terms"]');
+            // checkbox is checked
+            const defaultChecked = await termsCheckbox.evaluate((f) => f.checked);
+            assert.strictEqual(defaultChecked, fieldsWithDefault["terms"]);
+            // also the value = string
+            const defaultValue = await termsCheckbox.evaluate((f) => f.value);
+            assert.strictEqual(defaultValue, fieldsWithDefault["terms"].toString());
+        });
+
+        it("Check if changing the field value actually overwrites the default value", async function () {
+            const updatedFields = {
+                country: "USA",
+                "select-dropdown": "option 3",
+            };
+
+            await setInputValues(page, [{ name: "country", value: updatedFields["country"] }]);
+            await setSelectDropdownValue(page, "select", updatedFields["select-dropdown"]);
+
+            // input field default value
+            const countryInput = await getInputField(page, "country");
+            const updatedCountry = await countryInput.evaluate((f) => f.value);
+            assert.strictEqual(updatedCountry, updatedFields["country"]);
+
+            // dropdown default value is also getting set correctly
+            const selectDropdown = await waitForSTElement(page, "select");
+            const updatedOption = await selectDropdown.evaluate((f) => f.value);
+            assert.strictEqual(updatedOption, updatedFields["select-dropdown"]);
+        });
+
+        it("Check if default values are getting sent in signup-payload", async function () {
+            // directly submit the form and test the payload
+            const expectedDefautlValues = [
+                { id: "email", value: "test@one.com" },
+                { id: "password", value: "fakepassword123" },
+                { id: "terms", value: "true" },
+                { id: "select-dropdown", value: "option 2" },
+                { id: "country", value: "India" },
+            ];
+
+            let assertionError = null;
+            let interceptionPassed = false;
+
+            const requestHandler = async (request) => {
+                if (request.url().includes(SIGN_UP_API) && request.method() === "POST") {
+                    try {
+                        const postData = JSON.parse(request.postData());
+                        expectedDefautlValues.forEach(({ id, value }) => {
+                            let findFormData = postData.formFields.find((inputData) => inputData.id === id);
+                            if (findFormData) {
+                                assert.strictEqual(findFormData["value"], value, `Mismatch in payload for key: ${id}`);
+                            } else {
+                                throw new Error("Field not found in req.data");
+                            }
+                        });
+                        interceptionPassed = true;
+                        return request.respond({
+                            status: 200,
+                            headers: {
+                                "access-control-allow-origin": TEST_CLIENT_BASE_URL,
+                                "access-control-allow-credentials": "true",
+                            },
+                            body: JSON.stringify({
+                                status: "OK",
+                            }),
+                        });
+                    } catch (error) {
+                        assertionError = error; // Store the error
+                    }
+                }
+                return request.continue();
+            };
+
+            await page.setRequestInterception(true);
+            page.on("request", requestHandler);
+
+            try {
+                // Perform the button click and wait for all network activity to finish
+                await Promise.all([page.waitForNetworkIdle(), submitForm(page)]);
+            } finally {
+                page.off("request", requestHandler);
+                await page.setRequestInterception(false);
+            }
+
+            if (assertionError) {
+                throw assertionError;
+            }
+
+            if (!interceptionPassed) {
+                throw new Error("test failed");
+            }
+        });
+    });
+
+    describe("Incorrect field config test", function () {
+        beforeEach(async function () {
+            // set cookie and reload which loads the form fields with default values
+            await page.evaluate(() => window.localStorage.setItem("SHOW_INCORRECT_FIELDS", "YES"));
+
+            await page.reload({
+                waitUntil: "domcontentloaded",
+            });
+        });
+
+        it("Check if incorrect getDefaultValue throws error", async function () {
+            let pageErrorMessage = "";
+            page.on("pageerror", (err) => {
+                pageErrorMessage = err.message;
+            });
+
+            await page.reload({
+                waitUntil: "domcontentloaded",
+            });
+            await toggleSignInSignUp(page);
+
+            const expectedErrorMessage = "getDefaultValue for country must return a string";
+            assert(
+                pageErrorMessage.includes(expectedErrorMessage),
+                `Expected "${expectedErrorMessage}" to be included in page-error`
+            );
+        });
+
+        it("Check if non-string params to onChange throws error", async function () {
+            await page.evaluate(() => window.localStorage.setItem("INCORRECT_ONCHANGE", "YES"));
+            await page.reload({
+                waitUntil: "domcontentloaded",
+            });
+            await toggleSignInSignUp(page);
+
+            let pageErrorMessage = "";
+            page.on("pageerror", (err) => {
+                pageErrorMessage = err.message;
+            });
+
+            // check terms and condition checkbox since it emits non-string value => boolean
+            let termsCheckbox = await waitForSTElement(page, '[name="terms"]');
+            await page.evaluate((e) => e.click(), termsCheckbox);
+
+            const expectedErrorMessage = "terms value must be a string";
+            assert(
+                pageErrorMessage.includes(expectedErrorMessage),
+                `Expected "${expectedErrorMessage}" to be included in page-error`
+            );
         });
     });
 });