+++ test/helpers.js
@@ -417,6 +417,17 @@ export async function setInputValues(page, fields) {
     return await new Promise((r) => setTimeout(r, 300));
 }
 
+export async function setSelectDropdownValue(page, selector, optionValue) {
+    return await page.evaluate(
+        ({ selector, ST_ROOT_SELECTOR, optionValue }) => {
+            const select = document.querySelector(ST_ROOT_SELECTOR).shadowRoot.querySelector(selector);
+            select.value = optionValue;
+            select.dispatchEvent(new Event("change", { bubbles: true }));
+        },
+        { selector, ST_ROOT_SELECTOR, optionValue }
+    );
+}
+
 export async function clearBrowserCookiesWithoutAffectingConsole(page, console) {
     let toReturn = [...console];
     const client = await page.target().createCDPSession();